chore: split CI builds to new files
