+++ bin/ncu-ci
@@ -15,18 +15,20 @@ const {
   }
 } = require('../lib/ci/ci_type_parser');
 
+const { listBuilds } = require('../lib/ci/ci_utils');
+
+const { jobCache } = require('../lib/ci/build-types/job');
+const { PRBuild } = require('../lib/ci/build-types/pr_build');
+const { CommitBuild } = require('../lib/ci/build-types/commit_build');
+const { DailyBuild } = require('../lib/ci/build-types/daily_build');
+const { FailureAggregator } = require('../lib/ci/failure_aggregator');
+const { BenchmarkRun } = require('../lib/ci/build-types/benchmark_run');
+const { HealthBuild } = require('../lib/ci/build-types/health_build');
+const { CITGMBuild } = require('../lib/ci/build-types/citgm_build');
 const {
-  PRBuild,
-  BenchmarkRun,
-  CommitBuild,
-  CITGMBuild,
-  DailyBuild,
-  CITGMComparisonBuild,
-  HealthBuild,
-  listBuilds,
-  FailureAggregator,
-  jobCache
-} = require('../lib/ci/ci_result_parser');
+  CITGMComparisonBuild
+} = require('../lib/ci/build-types/citgm_comparison_build');
+
 const {
   RunPRJob
 } = require('../lib/ci/run_ci');
