+++ lib/ci/ci_utils.js
@@ -1,3 +1,22 @@
+const {
+  CI_DOMAIN,
+  parseJobFromURL,
+  CI_TYPES
+} = require('./ci_type_parser');
+
+const qs = require('querystring');
+
+const statusType = {
+  SUCCESS: 'SUCCESS',
+  FAILURE: 'FAILURE',
+  ABORTED: 'ABORTED',
+  UNSTABLE: 'UNSTABLE'
+};
+
+function getPath(url) {
+  return url.replace(`https://${CI_DOMAIN}/`, '').replace('api/json', '');
+}
+
 function getNodeName(url) {
   const re = /\/nodes=(.+?)\//;
   if (re.test(url)) {
@@ -17,16 +36,79 @@ function pad(any, length) {
   return (any + '').padEnd(length);
 }
 
-const statusType = {
-  SUCCESS: 'SUCCESS',
-  FAILURE: 'FAILURE',
-  ABORTED: 'ABORTED',
-  UNSTABLE: 'UNSTABLE'
-};
+function markdownRow(...args) {
+  let result = '';
+  for (const item of args) {
+    result += `| ${item} `;
+  }
+  return result + '|\n';
+}
+
+function filterBuild(builds, type) {
+  return builds
+    .filter(build => build.result === type)
+    .map(build => parseJobFromURL(build.url));
+}
+
+async function listBuilds(cli, request, type) {
+  // assert(type === COMMIT || type === PR)
+  const { jobName } = CI_TYPES.get(type);
+  const tree = 'builds[url,result]';
+  const url = `https://${CI_DOMAIN}/job/${jobName}/api/json?tree=${qs.escape(tree)}`;
+
+  cli.startSpinner(`Querying ${url}`);
+
+  const result = await request.json(url);
+  const builds = result.builds;
+  const failed = filterBuild(builds, statusType.FAILURE);
+  const aborted = filterBuild(builds, statusType.ABORTED);
+  const pending = filterBuild(builds, null);
+  const unstable = filterBuild(builds, statusType.UNSTABLE);
+  const success = filterBuild(builds, statusType.SUCCESS);
+  cli.stopSpinner('Done');
+
+  return {
+    success,
+    failed,
+    aborted,
+    pending,
+    unstable,
+    count: builds.length
+  };
+}
+
+function getHighlight(f) {
+  if (!f.reason) {
+    f.reason = 'failure not found';
+    return f.reason;
+  }
+  return f.reason.split('\n')[f.highlight]
+    .replace(/not ok \d+ /, '')
+    .replace(
+      /JNLP4-connect connection from \S+/, 'JNLP4-connect connection from ...'
+    )
+    .replace(/FATAL: Could not checkout \w+/, 'FATAL: Could not checkout ...')
+    .replace(
+      /error: pathspec .+ did not match any file\(s\) known to git/,
+      'error: pathspec ... did not match any file(s) known to git')
+    .replace(
+      /failed: no workspace for .+/,
+      'failed: no workspace for ...'
+    )
+    .replace(
+      /fatal: loose object \w+ \(stored in .git\/objects\/.+\) is corrupt/,
+      'fatal: loose object ... (stored in .git/objects/...) is corrupt')
+    .replace(/hudson\.plugins\.git\.GitException: /, '')
+    .replace(/java\.io\.IOException: /, '');
+}
 
 module.exports = {
   fold,
+  getHighlight,
   getNodeName,
+  getPath,
+  listBuilds,
+  markdownRow,
   pad,
   statusType
 };
