+++ lib/links.js
@@ -109,6 +109,10 @@ LinkParser.getPrURL = function({ owner, repo, prid }) {
   return `https://github.com/${owner}/${repo}/pull/${prid}`;
 };
 
+LinkParser.getMachineUrl = function(name) {
+  return `[${name}](https://ci.nodejs.org/computer/${name}/)`;
+};
+
 const PR_URL_RE = /PR-URL: https:\/\/github.com\/.+/;
 LinkParser.parsePrURL = function(text) {
   if (typeof text !== 'string') {
