+++ lib/pr_checker.js
@@ -24,7 +24,7 @@ const {
   isLiteCI,
   isFullCI
 } = require('./ci/ci_type_parser');
-const { PRBuild } = require('./ci/ci_result_parser');
+const { PRBuild } = require('./ci/build-types/pr_build');
 
 const GIT_CONFIG_GUIDE_URL = 'https://github.com/nodejs/node/blob/99b1ada/doc/guides/contributing/pull-requests.md#step-1-fork';
